do not set a maximum number of reconnection attempts
with fred's reconnection policy, once the max attempts is reached, it
will never try again. We actually want it to retry at some point, so we
use the exponential delay retry, along with a timeout so that
reconnecting does not block requests.
The maximum delay for reconnection is 2 seconds, so the router will
retry at this interval