add a warning if we think istio-proxy injection is causing problems (#3545)
We have encountered situations where the injection of istio-proxy in a
router pod (executing in kubernetes) causes strange networking errors
during uplink retrieval.

The root cause of the issue is that the router is executing and
attempting retrieve uplink schemas whilst the istio-proxy is modifying
network configuration at the same time.

This warning message will direct users to information which should help
them to configure their cluster or pod to avoid this problem.

fixes: #3533

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    ~~- [ ] Unit Tests~~
    ~~- [ ] Integration Tests~~
    - [x] Manual Tests

**Exceptions**

This is difficult to test, since the root cause (istio networking
re-configuration as the pod executes) is very difficult to reproduce in
a test environment.

Manual testing was performed by:
 - triggering a nightly build
 - deploying the resulting image to a test cluster
- using the istio pod annotation: `proxy.istio.io/config: '{
"holdApplicationUntilProxyStarts": false }'` to ensure the error
occurred
 - observing the desired warning message is produced

Here's some sample output from the test:
```
{"timestamp":"2023-08-07T13:04:15.762898Z","level":"WARN","message":"If your router is executing within a kubernetes pod, this failure may be caused by istio-proxy injection. See https://github.com/apollographql/router/issues/3533 for more details about how to solve this","target":"apollo_router::uplink"}
{"timestamp":"2023-08-07T13:04:15.782715Z","level":"ERROR","message":"fetch failed from all endpoints","target":"apollo_router::router::event::schema"}
```

 **Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

---------

Co-authored-by: Maria Elisabeth Schreiber <maria.schreiber@apollographql.com>