Require the main (GraphQL) server to shutdown before other servers
We'd like to have more control over the sequencing of server shutdown
so that the main (GraphQL) server is required to complete before other
servers are shutdown.

This is particularly desirable when the router is executing in
Kubernetes, since continuing to report live/ready checks during shutdown
is a requirement.

fixes: #3521