Update Gitleaks ignore file (#3546)
This updates the .gitleaks.toml file in the root of the repo. This file
is used to adjust gitleaks configuration when running against this repo.
Primarily, this configuration is used to create a repo-local allowlist
of detected "secret" values that should be allowed to remain in git
history. Usually, this happens if the detected value is not actually a
secret or if the detected value was a secret that has since been
revoked/rotated.