work on streams of Bytes for router requests
The Router level requests and responses have a stream of bytes as body.
Instead of accumulating the entire body, which might cause issues (DoS
on the request side, breaking `@defer` and subscriptions on the response
side), we will call the rhai code on every chunk.

On the request side:
- the first call contains the headers, but no body. This is done to
  avoid awaiting on the first chunk, which would need an async checkpoint,
  and having a clonable service as inner service for rhai
- subsequent calls will each contain a chunk, returned as a Vec<u8>, and
  cannot access the headers

On the response side:
- the first call contains the headers and the first chunk: so far we
  guarantee that there's at least one GraphQL response when creating the
  router response
- subsequent calls will each contain a chunk, returned as a Vec<u8>, and
  cannot access the headers