add lifecycle support to helm chart (#3570)
In order to reduce the instances of 502s during deployments I would like
to add a preStop hook with a sleep to give sending services enough time
to receive DNS updates. This PR exposes the container's lifecycle object
to accomplish that.

fixes #3563 

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [X] Changes are compatible[^1]
- [X] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [X] Manual Tests

**Exceptions**
- unit tests
- integration tests

Helm changes not applicable to unit/integration tests. Installed the
chart into a kind cluster with and without the property. Both worked.

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

---------

Co-authored-by: Gary Pennington <gary@apollographql.com>