Tweak metrics for Rust/JS GraphQL validation mismatch
Instead of having 4 different counters for the different kinds of
mismatches, and nothing for correct results, this PR uses two counters
(one for schema and one for query validation). It tags the count with
`match` if the Rust and JS results are the same, otherwise with
`false_positive` or `false_negative`.

In theory query/schema could also be a tag instead of a separate
counter, I'm not sure what the best practice is on when to use tags vs
separate counters, would appreciate input.