move the SpanMetricsExporter to a tracing Layer
SpanMetricsExporter is an OpenTelemetry Exporter, which means that if it
is loaded, then the entire OpenTelemetry handling infrastructure is
loaded, and especially the part that allocates data for the entire list
of spans.
Unfortunately, the SpanMetricsExporter was always loaded, even if we do
not export the metrics outside of the router, which means that there's a
constant overhead of telemetry even when it is not used.
This moves the SpanMetricsExporter to a lightweight tracing-subscriber
Layer which performs the same busy/idle accounting as
OpenTelemetryLayer, then generates the same events as before when the
span closes