Fix memory leak caused by Arc circular reference in `Notify` (#3692)
The issue is described in detail here:
https://github.com/apollographql/router/issues/3686

This PR uses `Weak` references to `Configuration` inside `Notify`'s
broadcast channel. This was causing a memory leak on router reload.

If the `upgrade` in the `subscription_task` is not successful, we cannot
update, so it is simply skipped this time, waiting for the next valid
configuration broadcast.

I'm not sure what the best way to test this is within the router test
suite. Happy to hear suggestions. I've validated it remediates the
memory leak issue on our end.

Co-authored-by: Jesse Rosenberger <git@jro.cc>