Do not record a trace if telemetry is not configured (#2999)
This introduces a `SamplingFilter` that wraps `OpenTelemetryLayer`. The layer has an overhead on every request, because it records data for each span, even if no exporters are set up. The filter handles sampling ahead of the layer, only sending a trace to the layer when it is actually needed, ie when it is sampled, and an exporter was configured.
This also reduces the overhead of sampling, by managing it outside of the OpenTelemetryLayer.

It is configured through a sampling ratio stored in an atomic u64, that is modifed when the telemetry configuration is activated