Add apollo.router.id to otlp metrics metadata (#3764)
Currently apollo metrics may have the same metadata between multiple
routers. This means that we can't disambiguate when multiple routers are
sending data.

Add a random UUID to the metadata.

This has no user facing impact.


Note: there is no unit test or integration test for this. We need to
take a broader look at how we do telemetry testing, and the solution
should probably involve otel collector. However that is beyond the scope
of this ticket.


<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

---------

Co-authored-by: bryn <bryn@apollographql.com>
Co-authored-by: Brandt Kinzle <114622981+bkinzle@users.noreply.github.com>