randomize the query planner cache warm up
When using query plan cache warm up, we take the list of most used
queries from the in memory cache, and plan them. The list will come in
order, from the most used to the least used.
If multipe router instances are using cache warm up, and are also
configured to use the distributed cache for query plans, then they would
all plan the list of queries in the same order, since they will have
roughly the same distribution in the in memory cache.
By randomizing the list of queries, we increase the chances that other
instances can get plans from the distributed cache during warm up
instead of calculating them, because other instances might have already
done the work.