Activate query planner cache warm up by default
Cache warm up has been in testing for long enough, and has proven highly
ezffective in smoothing out latencies when updating a schema. So this
will activate it by default. It can still be deactivated by configuring
it to 0:

```yaml
supergraph:
  query_planning:
    warmed_up_queries: 0
```

By default, it will now warm up the cache with the 30% most used queries