chore: Fix formatting in `pull_request_template.md` (#3814)
This has looked pretty funny to me for a while now and the fix was as
simple as using `:` instead of `.`! Inspired by @nathanmarcos doing it
as a one-off in https://github.com/apollographql/router/pull/3805. 😸

I also took the opportunity to add a nice horizontal rule that improves
the ascetics of the rendered result and made some other small tweaks to
text/formatting. Most notably though, I removed the suggestion to open a
issue labeled `manual test`. It takes up a lot of real-estate in the
text, and not a single issue has been opened or labeled as such. (Nor
does the label even exist. 😅 )