Introduce OneShotAsyncCheckpoint
The existing AsynCheckpoint requires `Clone` and thus introduces a need
for Service Buffering that reduces the performance and resiliance of the
router.

This new service set of Services, Layers and utility functions removes
the requirement for `Clone` and thus the requirement for service
buffering.

Existing uses of AsyncCheckpoint within the router are replaced with
OneShotAsyncCheckpoint along with the requirement to `buffer()` such
services.