Preserve large numbers in query transformations
In GraphQL source text, you can write an integer outside the int32 range
if a float is expected. apollo-rs parses these numbers as
`hir::Value::Int` because at that time, the expected type of the value is
not known, especially because it's possible to parse queries without
knowledge of the schema.

In the defer label transformation, any apollo-rs `hir::Value::Int` was
treated as an i32. This would cause an error if it's one such large integer.
The fully proper way to convert `hir::Value` to another type (ie. encoder or
serde_json, as is done elsewhere) would be to match on the actual
GraphQL type expected by the schema. But the simpler thing to do is to
just use a Float if the value is too big for an Int. This will
round-trip correctly.