Additional benign positive allowlisting (#3816)
This implements additional allowlisting for benign positive "secret"
detection. The values allowlisted in this PR are test/sample JWTs and
private keys.

Apologies for repeated PRs on this. These detections appeared after my
last PR for allowlisting secrets. I believe that gitleaks (the project
we're using to scan for secrets) updated their signatures to better
detect JWTs, causing more detections. I've run gitleaks on this repo by
hand and after this change, there are currently no additional detections
generated.

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`