Introduce OneShotAsyncCheckpoint (#3819)
The existing AsynCheckpoint requires `Clone` and thus introduces a need
for Service Buffering that reduces the performance and resiliance of the
router.

This new service set of Services, Layers and utility functions removes
the requirement for `Clone` and thus the requirement for service
buffering.

Existing uses of AsyncCheckpoint within the router are replaced with
OneShotAsyncCheckpoint along with the requirement to `buffer()` such
services.

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [x] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.