Fix Studio reporting when it is not configured
The sampling filter introduced in
b58a67c92d65b00c6c11c246e220dce9e2755de1 was making sure that the Apollo
exporter was configured when the key is present, but not in the right
place: that configuration is used in various places to set up
information in the spans and the context, and that is used to create the
Studio report and injecting the traces there.
In previous versions, the code was relying on the inject_schema_id
function in router_factory.rs to always insert a configuration for the
Apollo exporter.
This commit adds back the code that creates that configuration, but only
if the Apollo key and graph reference are present