persisted queries: improve Uplink failover and error handling (#3863)
If an Uplink request fails, the `stream_from_uplink` function
automatically tries the next configured Uplink endpoint. However, in the
case of the persisted queries feature, the Uplink response contains an
URL located in the same cloud provider as the Uplink endpoint (which is
used to fetch the full PQ list chunk). Before this PR, the PQ code
fetched this second URL "outside" of `stream_from_uplink`, so an error
downloading this file would *not* result in re-trying the Uplink request
itself on the next endpoint.

This PR factors out the body of `stream_from_uplink` into
`stream_from_uplink_transforming_new_response`, which takes an async
function that is called on any `UplinkResponse::New` to transform the
response. (`stream_from_uplink` passes the identity function here, so
its API is not affected.) If this function returns `Err`, the uplink
code moves on to the next endpoint URL, just like if the initial Uplink
fetch had failed. The PQ layer now fetches the PQ manifest chunk bodies
inside this async function instead of by mapping the uplink stream.

This means that if our GCP Uplink server is up but the GCS servers that
serve manifests from GCP are down, we will fail over to the AWS Uplink
server instead of just failing to fetch PQs.

Additionally, the response from Uplink can specify multiple valid URLs
for each chunk. (Uplink does not currently do this, but the protocol
allows for it.) Before this PR, Router would only look at the first URL
listed; after this PR, it tries each URL in order until finding one that
does not fail.

Additionally, before this PR, any errors fetching PQs from Uplink or GCS
after a successful startup were ignored. (They would be sent on the
`ready_sender` channel whose receiver is dropped, which could lead to a
`debug`-level logged message about how the channel wasn't open, but they
were not *directly* logged.) After this PR, we don't try to send these
errors on the channel after the channel was used once; instead, we log
them at the `error` level (similarly to how schema and license uplink
errors work). Note that this only occurs if Router fails to fetch PQs
from all configured Uplink endpoints; a failure to fetch from a single
endpoint is still only logged at the `debug` level.

Co-authored-by: Jeremy Lempereur <jeremy.lempereur@iomentum.com>

---------

Co-authored-by: o0Ignition0o <jeremy.lempereur@gmail.com>
Co-authored-by: Jeremy Lempereur <jeremy.lempereur@iomentum.com>