small experiment defer clones in query deduplication
the broadcast channel clones the message before sending it to every
recipient. Assuming the response is big, we could spend a lot of time
cloning and sending one by one, and the last requests in the list will
have increased latency. This is a test with storing the response in an
Arc, and cloning it in the context of the request, so that broadcasting
can be fast and cheap