Fix validation error with ID variable values overflowing i32 (#3896)
Fixes #3873

Input values for variables of type `ID` were previously validated as
"either like a GraphQL `Int` or like a GraphQL `String`". GraphQL `Int`
is specified as a signed 32-bit integer, such that values that overflow
fail validation. Applying this range restriction to `ID` values was
incorrect. Instead, validation for `ID` now accepts any JSON integer or
JSON string value, so that IDs larger than 32 bits can be used.

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [x] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [x] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.