re-write apollo_reports tests
Trying to make them work with the new batching features. I've simplified
the tests a lot to remove a number of globals, but there are still
things I don't like.

The main problem is the TEST lock which makes it so that each test
executes in series. It's annoying, but I can't figure out how to make
them work properly without that right now.