Update to otel 0.20.0 (#3649)
Updates Otel to 0.20

Otel 0.20 had a major change in the way that metrics were implemented
resulting in us having to do a complete overhaul on our integration.

Problems include:
* Some metrics traits were converted to structs, preventing the use of
our wrappers.
* It is no longer possible to migrate metrics from one prometheus
registry to another. This resulted in a significant amount of
restructuring and special handling to allow prometheus meter provider to
remain across reloads.
* Related to the above many tests were broken due to their reliance on
prometheus and our lack of testability.

Unfortunately the above meant a significant amount of rework just to
retain the equivalent tests that we had before.

Some dev docs have been created to try and give a higher level overview
of how things fit together:

https://github.com/apollographql/router/blob/bryn/otel-update/dev-docs/metrics.md

Perf looks like it hasn't regressed.


<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

---------

Signed-off-by: Benjamin Coenen <5719034+bnjjj@users.noreply.github.com>
Co-authored-by: bryn <bryn@apollographql.com>
Co-authored-by: Coenen Benjamin <benjamin.coenen@hotmail.com>