Metrics and root span creation now happen first
Our metrics are currently mutated at the supergraph service. This is a throwback from when the router service did not exist.
Move operation count metrics to the first thing that happens in the pipeline.
Note that this won't catch disconnects because they are reliant on the status code of the response, which will stop processing before the metric is incremented.

Fixes #3915