rework apollo_reports testing further
It's really hard to test report generation for metrics. Traces at least
are aggregated within a span, so no special handling is required. The
existing metrics tests only send in 1 request, so they work by accident.

The batch metric test generates two requests (batch of 2) and so this
makes the existing testing mechanisms very sad. I've decided on a
compromise for batch testing. Just check that the number of requests
sums to 2 across all of the metrics reports which are generated.

Not ideal, but something we could try to improve in the future.