experimental support for query batching (#3837)
Things needing attention

 - [x] Configuration to match the design
- [x] Decide if cloning `parts` is acceptable (in review) It isn't, so
not cloned but re-built.
- [x] Decide if cloning context for each `SupergraphRequest` is
acceptable (in review)
- [x] Should we do special handling if `@defer` or `subscription`
detected in batch?
 - [x] Metrics
- [x] Find someone happy to test from an appropriately configured Apollo
Client and verify it works
- [x] Modify Apollo telemetry to create separate root traces for each
batch entry
 
Fixes #126

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [x] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [x] Integration Tests
    - [x] Manual Tests

**Exceptions**

Manual testing was performed with `curl` and [apollo
client](https://www.apollographql.com/docs/react/api/link/apollo-link-batch-http/)

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.

---------

Co-authored-by: Edward Huang <edward.huang@apollographql.com>
Co-authored-by: Geoffroy Couprie <apollo@geoffroycouprie.com>
Co-authored-by: Maria Elisabeth Schreiber <maria.schreiber@apollographql.com>